requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
Authlib==1.5.2
babel==2.17.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
google-auth==2.39.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
text-unidecode==1.3
time-machine==2.16.0
tortoise-orm==0.24.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
